Split up lib.rs into files
